EPMLSTRTAI-359: Rename package for integration tests. Move JavaFileWriter test to it package
